requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-qdrant
qdrant-client
sentence-transformers
pypdf
docx2txt
pydantic-settings
python-dotenv
flake8
black
isort
mypy
pytest
httpx
